gps: Introduce verify subpackage
This is a start at isolating verification logic into a discrete package.
Not sure how far we'll be able to make this go without creating some
import loops.
